requirements.txt
alembic==1.8.0
chess==1.9.2
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
greenlet==1.1.2
gunicorn==20.0.4
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pytz==2022.1
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
